+++ src/modules/mo-project-dto/enums/index.ts
@@ -24,3 +24,4 @@ export * from './xterm-cluster-operator-log-component-type.enum';
 export * from './project-namespace-service-list-cronjob-job-tile-status.enum';
 export * from './project-namespace-service-list-cronjob-job-status.enum';
 export * from './project-notification-settings-type.enum';
+export * from './project-namespace-service-envvar-vault-type.enum';