+++ src/modules/mo-project-dto/enums/project-namespace-service-envvar-vault-type.enum.ts
@@ -1,0 +1,4 @@
+export enum ProjectNamespaceServiceEnvVarVaultTypeEnum {
+  MOGENIUS_VAULT = 'MOGENIUS_VAULT',
+  HASHICORP_EXTERNAL_VAULT = 'HASHICORP_EXTERNAL_VAULT'
+}